refactor: optimise code, schema and meta added
