+++ projects/commudle-admin/src/app/feature-modules/public-blogs/components/blogs-list/blogs-list.component.html
