+++ projects/commudle-admin/src/app/feature-modules/public-blogs/public-blogs-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { BlogComponent } from './components/blog/blog.component';
-import { BlogsListComponent } from './components/blogs/blogs-list.component';
+import { BlogsListComponent } from './components/blogs-list/blogs-list.component';
 
 const routes: Routes = [
   {
