+++ projects/commudle-admin/src/app/feature-modules/public-blogs/public-blogs.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { BlogsListComponent } from './components/blogs/blogs-list.component';
+import { BlogsListComponent } from './components/blogs-list/blogs-list.component';
 import { PublicBlogsRoutingModule } from './public-blogs-routing.module';
 import { NbEvaIconsModule } from '@nebular/eva-icons';
 import { NbButtonModule, NbCardModule, NbIconModule, NbSpinnerModule } from '@nebular/theme';
